build fixes, moved and renamed designer dialog
